feat: analytics
