+++ src/app/(platform)/dashboard/@admin/opportunities/page.js
@@ -33,3 +33,7 @@ const AdminOpportunities = async () => {
 };
 
 export default AdminOpportunities;
+
+export const metadata = {
+  title: "Admin Opportunities",
+};
