+++ src/app/(platform)/dashboard/@admin/users/page.js
@@ -26,7 +26,7 @@ const AdminUsers = async () => {
   const users = await getUsers();
 
   return (
-    <Container size="6xl" className="mt-10 text-center">
+    <Container size="6xl" className="text-center space-y-4 mt-10">
       <Heading>Users</Heading>
       <AdminUsersTable users={users} />
     </Container>
@@ -34,3 +34,7 @@ const AdminUsers = async () => {
 };
 
 export default AdminUsers;
+
+export const metadata = {
+  title: "Admin Users",
+};
