+++ src/app/(platform)/profile/page.js
@@ -3,3 +3,7 @@ import Profile from "@/components/Profile";
 const Page = () => <Profile />;
 
 export default Page;
+
+export const metadata = {
+  title: "Profile",
+};
